+++ Source/GateRimSG1/GateRimSG1Bootstrap.cs
@@ -1,3 +1,4 @@
+using GateRimSG1.Goauld;
 using Verse;
 
 namespace GateRimSG1
@@ -7,6 +8,8 @@ namespace GateRimSG1
     {
         static GateRimSG1Bootstrap()
         {
+            GateRimSG1HarmonyInitializer.Apply();
+
             GR_Log.Message(
                 $"Version {typeof(GateRimSG1Bootstrap).Assembly.GetName().Version} loaded.");
         }