+++ Source/GateRimSG1/Goauld/IncidentWorker_TokraTherapeuticOpportunity.cs
@@ -32,6 +32,11 @@ namespace GateRimSG1.Goauld
 
         protected override bool CanFireNowSub(IncidentParms parms)
         {
+            if (!TokraFactionUtility.HasPersistentFaction())
+            {
+                return false;
+            }
+
             if (!base.CanFireNowSub(parms))
             {
                 return false;